move MessageTask to a separate file
